Refactoring updated code to remove compilation errors after moving xml files.
