+++ SemanticDataSolution/UANodeSetValidation/AddressSpaceContext.cs
@@ -1,10 +1,8 @@
 ﻿
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using System.Text;
+using UAOOI.SemanticData.UANodeSetValidation.XML;
 
 namespace UAOOI.SemanticData.UANodeSetValidation
 {
