+++ SemanticDataSolution/UANodeSetValidation/XML/UANodeSet.Utils.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace UAOOI.SemanticData.UANodeSetValidation
+namespace UAOOI.SemanticData.UANodeSetValidation.XML
 {
   public partial class UANodeSet
   {
